Split Reachability into two implementations.
