+++ src/bindings/algorithms/reachability/mod.rs
@@ -8,6 +8,7 @@ mod reachability_config;
 mod reachability_config_python;
 mod reachability_error;
 mod reachability_impl;
+mod reachablility_impl_python;
 
 pub use reachability_config::ReachabilityConfig;
 pub use reachability_error::ReachabilityError;
